[surface model] Static cast where needed
This MR prevents a bunch of implicit casts from double to float, which otherwise spam warnings when compiling with `-Wdouble-promotion`.
All implicit double promotion in the surface model when running in double precision are fixed, except those from printing surfaces.
For mixed precision, there is still one transformation that is taken directly in the solid model (in double precision) which leads to unwanted promotion. This will be fixed in another MR.
Also, the Arb4 and Torus are not float promotion free, but they need to be assessed separately anyway, since the torus could never be stable in single precision.

A new stricter tolerance in single precision was added, so that I could replace all compile-time `kTolerance` without breaking mixed precision by having a way too high tolerance.

No speedup was measurable with the change (despite fixing a matrix multiplication), the results are unchanged.

However, this cleans up the compilation with this flag that could help finding performance-critical, unwanted promotions from float to double. There are quite a few warnings left (mostly in the solid model) and they are left for future cleaning.

* fix double promotion for mixed precision

* fix clang formatting breaking ArgParser

* static cast where needed

+++ test/benchmark/ArgParser.h
@@ -15,7 +15,7 @@ double getDoubleOpt(char **begin, char **end, const std::string &option, double 
   return defaultval;
 }
 
-double getIntOpt(char **begin, char **end, const std::string &option, int defaultval)
+int getIntOpt(char **begin, char **end, const std::string &option, int defaultval)
 {
   char **itr = std::find(begin, end, option);
   if (itr != end && ++itr != end) {
